app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-socketio==5.11.4
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
aiofiles==24.1.0